requirements.txt
yfinance>=0.2.31
requests-cache>=1.1.0
requests-ratelimiter>=0.4.0
pandas>=2.0.0
numpy>=1.24.0
google-genai>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
